refactor(api): extract decideApprovalRequest from the approvals decide route (behavior-preserving) (#4189)
`decideHandler`'s body (everything after the `id` param read) moves verbatim
into `services/approvals/decideApprovalRequest.ts` as
`decideApprovalRequest(input): Promise<{ httpStatus, body }>`. Every
`return c.json(X, N)` becomes `return { httpStatus: N, body: X }`;
`c.get('auth')` becomes `input.auth`; `c.req.param('id')` becomes
`input.id`. The two routes are now thin adapters over it.

The row DTO (`serialize`) and its intent projections (`toIntentTargetRef`,
`toIntentAttribution`, `resolveTargetDevices`) move with it and are
re-imported by the route, so the inbox list and the decide response cannot
drift apart — and so the upcoming batch path can hand back the same per-row
bodies.

Proof: `npx vitest run src/routes/approvals` — 105 tests, 0 test edits.
